Fix closest-to to actually work (and rename to equivalent-coords)
Previously, closest-to would decide whether it already knew about a place based on if the :places cache had any data, but, this was always true, because the API would search for the place (putting it in the cache) before computing (and querying darksky).

Fixed by changing closest-to to be based on if there is existing darsky data for a given lon/lat (which, thanks to the recent commit which changed the cache directory strcuture, is very fast).